Move FileOutput to helpers module
